refactor(objects): move object source traits down (#1100)
