+++ crates/objects/src/object/source.rs
@@ -1,13 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 //! Read-only object source traits for graph walkers.
 
-use super::Result;
-use crate::{
-    object::{Blob, ContentHash, State, StateId, Tree},
-    store::ObjectStore,
-};
+use crate::error::Result;
 
-/// Read-only subset of [`ObjectStore`] needed by object graph walkers.
+use super::{Blob, ContentHash, State, StateId, Tree};
+
+/// Read-only object access needed by object graph walkers.
 pub trait ObjectSource {
     fn get_tree(&self, hash: &ContentHash) -> Result<Option<Tree>>;
     fn get_state(&self, id: &StateId) -> Result<Option<State>>;
@@ -18,28 +16,6 @@ pub trait ObjectSource {
         Ok(self
             .get_blob(hash)?
             .map(|blob| bytes::Bytes::from(blob.into_content())))
-    }
-}
-
-impl<S: ObjectStore + ?Sized> ObjectSource for S {
-    #[inline]
-    fn get_tree(&self, hash: &ContentHash) -> Result<Option<Tree>> {
-        ObjectStore::get_tree(self, hash)
-    }
-
-    #[inline]
-    fn get_state(&self, id: &StateId) -> Result<Option<State>> {
-        ObjectStore::get_state(self, id)
-    }
-
-    #[inline]
-    fn get_blob(&self, hash: &ContentHash) -> Result<Option<Blob>> {
-        ObjectStore::get_blob(self, hash)
-    }
-
-    #[inline]
-    fn get_blob_bytes(&self, hash: &ContentHash) -> Result<Option<bytes::Bytes>> {
-        ObjectStore::get_blob_bytes(self, hash)
     }
 }
